+++ package.json
@@ -6,7 +6,8 @@
   "dependencies": {
     "@babel/generator": "^7.22.15",
     "@babel/parser": "^7.22.16",
-    "@babel/traverse": "^7.22.15",
+    "@babel/traverse": "^7.22.17",
+    "@babel/types": "^7.22.17",
     "@braintree/sanitize-url": "^6.0.2",
     "@keypom/core": "^1.0.2",
     "@monaco-editor/react": "^4.5.1",
@@ -18,13 +19,9 @@
     "@near-wallet-selector/near-wallet": "^8.1.0",
     "@near-wallet-selector/neth": "^8.1.0",
     "@near-wallet-selector/sender": "^8.1.0",
-    "@teleporthq/teleport-component-generator": "^0.42.0-alpha.0",
-    "@teleporthq/teleport-component-generator-react": "^0.40.0-alpha.0",
-    "@teleporthq/teleport-plugin-css": "^0.42.0-alpha.0",
-    "@teleporthq/teleport-plugin-react-styled-components": "^0.42.0-alpha.0",
-    "@teleporthq/teleport-postprocessor-prettier-js": "^0.42.0-alpha.0",
-    "@teleporthq/teleport-shared": "^0.42.0-alpha.0",
-    "@teleporthq/teleport-types": "^0.42.0-alpha.0",
+    "@teleporthq/teleport-component-generator-react": "^0.30.0",
+    "@teleporthq/teleport-plugin-react-styled-components": "^0.16.0",
+    "@teleporthq/teleport-types":"^0.30.0",
     "@web3-onboard/core": "^2.16.2",
     "@web3-onboard/injected-wallets": "^2.8.4",
     "@web3-onboard/ledger": "^2.4.4",
@@ -39,7 +36,7 @@
     "error-polyfill": "^0.1.2",
     "local-storage": "^2.0.0",
     "near-api-js": "^2.1.3",
-    "near-social-vm": "./near-social-vm-v2.2.4.tgz",
+    "near-social-vm": "./VM-2.4.0.tar.gz",
     "normalize.css": "^8.0.1",
     "prettier": "^2.7.1",
     "react": "^18.2.0",
