--- src/components/custom/react-mapping.json
@@ -1,7 +1,0 @@
-{
-  "elements": {
-    "img": {
-      "elementType": "widget"
-    }
-  }
-}
